Update component name
